separate filesystem and menu interface
